Added the user's specific routes to the blueprint
