+++ enigma/routes.py
@@ -19,72 +19,10 @@ def home():
 def about():
     return render_template('about.html', title="About")
 
-@app.route("/register", methods=["GET", "POST"])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        hashed_password = argon2.generate_password_hash(form.password.data)
-        user = User(username=form.username.data, email=form.email.data, password=hashed_password)
-        db.session.add(user)
-        db.session.commit()
-        flash("Your account has been created! You are now able to log in.", "success")
-        return redirect(url_for("login"))
-    return render_template("register.html", title="Register", form=form)
 
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and argon2.check_password_hash(user.password, form.password.data):
-            login_user(user, remember=form.remember.data)
-            next_page = request.args.get("next")
-            return redirect(next_page) if next_page else redirect(url_for("home"))    
-        else:
-            flash("Login unsuccessful. Please check email and password.", "danger")
-    return render_template("login.html", title="Login", form=form)
 
-@app.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for("home"))
 
-def save_picture(form_picture):
-    random_hex = secrets.token_hex(8)
-    _, f_ext = os.path.splitext(form_picture.filename)
-    picture_fn = random_hex + f_ext
-    picture_path = os.path.join(app.root_path, "static/profile_pics", picture_fn)
 
-    # Resize image
-    output_size = (125, 125)
-    i = Image.open(form_picture)
-    i.thumbnail(output_size)
-    i.save(picture_path)
-
-    return picture_fn
-
-@app.route("/account", methods=["GET", "POST"])
-@login_required
-def account():
-    form = UpdateAccountForm()
-    if form.validate_on_submit():
-        if form.picture.data:
-            picture_file = save_picture(form.picture.data)
-            current_user.image_file = picture_file
-        current_user.username = form.username.data
-        current_user.email = form.email.data
-        db.session.commit()
-        flash("You account has been updated!", "success")
-        return redirect(url_for("account"))
-    elif request.method == "GET":
-        form.username.data = current_user.username
-        form.email.data = current_user.email
-    image_file = url_for("static", filename=f"profile_pics/{current_user.image_file}")
-    return render_template("account.html", title="Account", image_file=image_file, form=form)
 
 @app.route("/post/new", methods=["GET", "POST"])
 @login_required
@@ -132,51 +70,4 @@ def delete_post(post_id):
     flash("Your post has been deleted!", "success")
     return redirect(url_for("home"))
 
-@app.route("/user/<string:username>")
-def user_posts(username):
-    page = request.args.get("page", 1, type=int)
-    user = User.query.filter_by(username=username).first_or_404()
-    posts = Post.query.filter_by(author=user).order_by(Post.date_posted.desc()).paginate(page=page, per_page=5)
-    return render_template('user_posts.html', posts=posts, user=user)
 
-def send_reset_email(user):
-    token = user.get_reset_token()
-    msg = Message("Password Reset Request", sender="noreply@enigma.com", recipients=[user.email])
-    msg.body = f'''To reset your password, visit the following link:
-{url_for('reset_token', token=token, _external=True)}
-
-If you did not make this request then simply ignore this email and no changes will be made.
-'''
-    # We do not have a service for that at the moment.
-    # mail.send(msg)
-    # For now we print the token in the console.
-    print(url_for('reset_token', token=token, _external=True))
-
-@app.route("/reset_password", methods=["GET", "POST"])
-def reset_request():
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    form = RequestResetForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user: send_reset_email(user)
-        flash("An email has been sent with instructions to rest your password.", "info")
-        return redirect(url_for("login"))
-    return render_template('reset_request.html', title="Reset Password", form=form)
-
-@app.route("/reset_password/<token>", methods=["GET", "POST"])
-def reset_token(token):
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    user = User.verify_reset_token(token)
-    if user is None:
-        flash("That is an invalid or expired token.", "warning")
-        return redirect(url_for("reset_request"))
-    form = ResetPasswordForm()
-    if form.validate_on_submit():
-        hashed_password = argon2.generate_password_hash(form.password.data)
-        user.password = hashed_password
-        db.session.commit()
-        flash("Your password has been updated! You are now able to log in.", "success")
-        return redirect(url_for("login"))
-    return render_template('reset_token.html', title="Reset Password", form=form)
